tests: migrate getRepoUrl
